requirements.txt
pygame>=2.0.0
websockets>=10.0 
websocket-client>=1.0.0 